Move image capture classes into src/capture/image
