+++ src/capture/image/image_capturer.cpp
