+++ src/capture/image/image_decoder.cpp
