+++ src/capture/image/image_decoder.h
