+++ src/capture/image/image_scaler.cpp
