+++ src/capture/image/image_scaler.h
