+++ src/capture/image/png_writer.cpp
