+++ src/capture/image/png_writer.h
