+++ vs/dosbox.vcxproj
@@ -513,11 +513,11 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClCompile Include="..\src\capture\capture_audio.cpp" />
     <ClCompile Include="..\src\capture\capture_midi.cpp" />
     <ClCompile Include="..\src\capture\capture_video.cpp" />
-    <ClCompile Include="..\src\capture\image_capturer.cpp" />
-    <ClCompile Include="..\src\capture\image_decoder.cpp" />
-    <ClCompile Include="..\src\capture\image_saver.cpp" />
-    <ClCompile Include="..\src\capture\image_scaler.cpp" />
-    <ClCompile Include="..\src\capture\png_writer.cpp" />
+    <ClCompile Include="..\src\capture\image\image_capturer.cpp" />
+    <ClCompile Include="..\src\capture\image\image_decoder.cpp" />
+    <ClCompile Include="..\src\capture\image\image_saver.cpp" />
+    <ClCompile Include="..\src\capture\image\image_scaler.cpp" />
+    <ClCompile Include="..\src\capture\image\png_writer.cpp" />
     <ClCompile Include="..\src\cpu\callback.cpp" />
     <ClCompile Include="..\src\cpu\core_dynrec.cpp" />
     <ClCompile Include="..\src\cpu\core_dyn_x86.cpp" />
@@ -798,11 +798,11 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClInclude Include="..\src\capture\capture_audio.h" />
     <ClInclude Include="..\src\capture\capture_midi.h" />
     <ClInclude Include="..\src\capture\capture_video.h" />
-    <ClInclude Include="..\src\capture\image_capturer.h" />
-    <ClInclude Include="..\src\capture\image_decoder.h" />
-    <ClInclude Include="..\src\capture\image_saver.h" />
-    <ClInclude Include="..\src\capture\image_scaler.h" />
-    <ClInclude Include="..\src\capture\png_writer.h" />
+    <ClInclude Include="..\src\capture\image\image_capturer.h" />
+    <ClInclude Include="..\src\capture\image\image_decoder.h" />
+    <ClInclude Include="..\src\capture\image\image_saver.h" />
+    <ClInclude Include="..\src\capture\image\image_scaler.h" />
+    <ClInclude Include="..\src\capture\image\png_writer.h" />
     <ClInclude Include="..\src\cpu\core_dynrec\decoder.h" />
     <ClInclude Include="..\src\cpu\core_dynrec\decoder_basic.h" />
     <ClInclude Include="..\src\cpu\core_dynrec\decoder_opcodes.h" />
