+++ vs/dosbox.vcxproj.filters
@@ -13,20 +13,20 @@
     <ClCompile Include="..\src\capture\capture_video.cpp">
       <Filter>src\capture</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\capture\image_capturer.cpp">
-      <Filter>src\capture</Filter>
+    <ClCompile Include="..\src\capture\image\image_capturer.cpp">
+      <Filter>src\capture\image</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\capture\image_decoder.cpp">
-      <Filter>src\capture</Filter>
+    <ClCompile Include="..\src\capture\image\image_decoder.cpp">
+      <Filter>src\capture\image</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\capture\image_saver.cpp">
-      <Filter>src\capture</Filter>
+    <ClCompile Include="..\src\capture\image\image_saver.cpp">
+      <Filter>src\capture\image</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\capture\image_scaler.cpp">
-      <Filter>src\capture</Filter>
+    <ClCompile Include="..\src\capture\image\image_scaler.cpp">
+      <Filter>src\capture\image</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\capture\png_writer.cpp">
-      <Filter>src\capture</Filter>
+    <ClCompile Include="..\src\capture\image\png_writer.cpp">
+      <Filter>src\capture\image</Filter>
     </ClCompile>
     <ClCompile Include="..\src\cpu\callback.cpp">
       <Filter>src\cpu</Filter>
@@ -996,20 +996,20 @@
     <ClInclude Include="..\src\capture\capture_video.h">
       <Filter>src\capture</Filter>
     </ClInclude>
-    <ClInclude Include="..\src\capture\image_capturer.h">
-      <Filter>src\capture</Filter>
+    <ClInclude Include="..\src\capture\image\image_capturer.h">
+      <Filter>src\capture\image</Filter>
     </ClInclude>
-    <ClInclude Include="..\src\capture\image_decoder.h">
-      <Filter>src\capture</Filter>
+    <ClInclude Include="..\src\capture\image\image_decoder.h">
+      <Filter>src\capture\image</Filter>
     </ClInclude>
-    <ClCompile Include="..\src\capture\image_saver.h">
-      <Filter>src\capture</Filter>
+    <ClCompile Include="..\src\capture\image\image_saver.h">
+      <Filter>src\capture\image</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\capture\image_scaler.h">
-      <Filter>src\capture</Filter>
+    <ClCompile Include="..\src\capture\image\image_scaler.h">
+      <Filter>src\capture\image</Filter>
     </ClCompile>
-    <ClInclude Include="..\src\capture\png_writer.h">
-      <Filter>src\capture</Filter>
+    <ClInclude Include="..\src\capture\image\png_writer.h">
+      <Filter>src\capture\image</Filter>
     </ClInclude>
     <ClInclude Include="..\src\cpu\core_dynrec\decoder.h">
       <Filter>src\cpu\core_dynrec</Filter>
@@ -1449,6 +1449,9 @@
     </Filter>
     <Filter Include="src\capture">
       <UniqueIdentifier>{e19191a3-0830-4c33-b6f4-41dd6f76c211}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="src\capture\image">
+      <UniqueIdentifier>{429c5b2d-99ce-4d8d-8eea-7702f851d8ee}</UniqueIdentifier>
     </Filter>
     <Filter Include="src\cpu">
       <UniqueIdentifier>{034c69b6-a49d-4c92-b15c-9939085f12d7}</UniqueIdentifier>
